# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StoryForge MVP API
  version: 0.3.0
  description: MVP API — avatar upload (JPG/PNG ≤5MB, thumb 200²), Google OAuth (JWT sf_session), admin moderation, chat SSE (heartbeat + timeout), snapshots, stories, branches export annotate, profile, account deletion (§2.3), moderation, export, feed (author_display for deleted users). Production mutations on /api/* Origin-guard (middleware). Admins STORYFORGE_ADMIN_USER_IDS.
servers:
  - url: http://localhost:3000
paths:
...
  /api/account/delete:
    post:
      summary: Delete account — requires literal confirm text and works_action; clears session cookies on success
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm, works_action]
              properties:
                confirm:
                  type: string
                  enum: [确认注销]
                works_action:
                  type: string
                  enum: [anonymize_published, remove_all]
                  description: anonymize_published keeps published content with author_display 已注销用户; remove_all deletes authored content
      responses:
        "200":
          description: Account deleted; sf_session / sf_user_id cleared
        "400":
          description: Invalid body, demo user, or validation error
        "410":
          description: Already deleted
        "429":
          description: Rate limited
...
  /api/follows/toggle:
    post:
      summary: Toggle follow (cannot follow deleted authors)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [author_id]
              properties:
                author_id:
                  type: string
      responses:
        "200":
          description: Success
        "400":
          description: Self-follow, deleted/missing author, etc.
...
  /api/admin/moderation/{id}:
    patch:
      summary: Approve, take-down target content, or reject with remark
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, taken_down, rejected]
                audit_remark:
                  type: string
      responses:
        "200":
          description: Success
  /api/profile:
    get:
      summary: Current user profile
      responses:
        "200":
          description: Success
        "410":
          description: Account deleted — session cookies cleared
    patch:
      summary: Update profile
      responses:
        "200":
          description: Success
        "410":
          description: Account deleted — session cookies cleared
...
  /api/favorites/toggle:
    post:
      summary: Toggle favorite on story / character / world
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_type, target_id]
              properties:
                target_type:
                  type: string
                  enum: [story, character, world]
                target_id:
                  type: string
      responses:
        "200":
          description: Success
...
  /api/stories/{id}/branches/{branchId}:
    patch:
      summary: Update branch (archive, retitle, reparent)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: branchId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
    delete:
      summary: Delete branch if no children
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: branchId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
...